Extract K8s API authentication into specialised client module
