+++ kopf/config.py
@@ -4,12 +4,9 @@ import concurrent.futures
 import logging
 from typing import Optional
 
-import click
 import kubernetes
 import kubernetes.client.rest
-import urllib3.exceptions
 
-logger = logging.getLogger(__name__)
 format = '[%(asctime)s] %(name)-20.20s [%(levelname)-8.8s] %(message)s'
 
 
@@ -21,42 +18,6 @@ LOGLEVEL_ERROR = 40
 """ Event loglevel to log only errors and critical events. """
 LOGLEVEL_CRITICAL = 50
 """ Event loglevel to log only critical events(basically - no events). """
-
-
-class LoginError(click.ClickException):
-    """ Raised when the operator cannot login to the API. """
-
-
-def login():
-    """
-    Login the the Kubernetes cluster, locally or remotely.
-    """
-
-    # Configure the default client credentials for all possible environments.
-    try:
-        kubernetes.config.load_incluster_config()  # cluster env vars
-        logger.debug("configured in cluster with service account")
-    except kubernetes.config.ConfigException as e1:
-        try:
-            kubernetes.config.load_kube_config()  # developer's config files
-            logger.debug("configured via kubeconfig file")
-        except kubernetes.config.ConfigException as e2:
-            raise LoginError(f"Cannot authenticate neither in-cluster, nor via kubeconfig.")
-
-    # Make a sample API call to ensure the login is successful,
-    # and convert some of the known exceptions to the CLI hints.
-    try:
-        api = kubernetes.client.CoreApi()
-        api.get_api_versions()
-    except urllib3.exceptions.HTTPError as e:
-        raise LoginError("Cannot connect to the Kubernetes API. "
-                         "Please configure the cluster access.")
-    except kubernetes.client.rest.ApiException as e:
-        if e.status == 401:
-            raise LoginError("Cannot authenticate to the Kubernetes API. "
-                             "Please login or configure the tokens.")
-        else:
-            raise
 
 
 def configure(debug=None, verbose=None, quiet=None):
