+++ tests/cli/test_help.py
@@ -1,7 +1,7 @@
 
 
 def test_help_in_root(invoke, mocker):
-    login = mocker.patch('kopf.config.login')
+    login = mocker.patch('kopf.clients.auth.login')
 
     result = invoke(['--help'])
 
@@ -15,7 +15,7 @@ def test_help_in_root(invoke, mocker):
 
 
 def test_help_in_subcommand(invoke, mocker):
-    login = mocker.patch('kopf.config.login')
+    login = mocker.patch('kopf.clients.auth.login')
     preload = mocker.patch('kopf.utilities.loaders.preload')
     real_run = mocker.patch('kopf.reactor.queueing.run')
