+++ tests/cli/test_login.py
@@ -1,7 +1,7 @@
 import kubernetes
 import pytest
 
-from kopf.config import login, LoginError
+from kopf.clients.auth import login, LoginError
 
 
 @pytest.fixture(autouse=True)
